# textparse: Use cache for protoparse labels, break interface.