--- a/model/textparse/promparse.go
+++ b/model/textparse/promparse.go
@@ -223,9 +223,9 @@ func (p *PromParser) Comment() []byte {
 	return p.text
 }
 
-// Metric writes the labels of the current sample into the passed labels.
+// Labels writes the labels of the current sample into the passed labels.
 // It returns the string from which the metric was parsed.
-func (p *PromParser) Metric(l *labels.Labels) string {
+func (p *PromParser) Labels(l *labels.Labels) {
 	// Copy the buffer to a string: this is only necessary for the return value.
 	s := string(p.series)
 
@@ -246,8 +246,6 @@ func (p *PromParser) Metric(l *labels.Labels) string {
 
 	p.builder.Sort()
 	*l = p.builder.Labels()
-
-	return s
 }
 
 // Exemplar implements the Parser interface. However, since the classic