Depends on https://github.com/prometheus/prometheus/pull/15731

This implements what's discussed in https://github.com/prometheus/prometheus/pull/15731#discussion_r1950946229

It's quite complex due to https://protobuf.dev/programming-guides/serialization-not-canonical/ and magic suffixes/labels for classic series, but seems doable.