@bwplotka 2025-02-12
/prombench main --bench.version=bench/protofirst