@bwplotka 2025-02-13
Something is off with the prombench results I expected bigger improvement. Might need some work, but pausing it given https://github.com/prometheus/prometheus/pull/15731 makes native histogram parsing good enough (on par really with other protocols), focusing on other things for the next weeks. 

Anybody can continue this work if they want!